EditProfile Folder Restructured
