+++ client/src/helpers/APICalls/editProfile.ts
@@ -1,0 +1,27 @@
+import { AuthApiData } from '../../interface/AuthApiData';
+import { FetchOptions } from '../../interface/FetchOptions';
+
+const register = async (
+  firstName: string,
+  lastName: string,
+  gender: string,
+  dateOfBirth: number,
+  email: string,
+  phone: number,
+  address: string,
+  description: string,
+): Promise<AuthApiData> => {
+  const fetchOptions: FetchOptions = {
+    method: 'POST',
+    headers: { 'Content-Type': 'application/json' },
+    body: JSON.stringify({ firstName, lastName, gender, dateOfBirth, email, phone, address, description }),
+    credentials: 'include',
+  };
+  return await fetch(`/auth/register`, fetchOptions)
+    .then((res) => res.json())
+    .catch(() => ({
+      error: { message: 'Unable to connect to server. Please try again' },
+    }));
+};
+
+export default register;
